physics-lib - Preparing the PR to upstream. Project CLang's configuration is upgraded to the 2b flag for the C++ standard. Internal partition for the details is a raw module
